+++ bin/run-selenium-desktop.sh
@@ -1,2 +1,2 @@
 #!/bin/sh
-java -Xms512m -Xmx2048m -Dwebdriver.chrome.driver="./bin/chromedriver" -jar ./bin/selenium-server-standalone.jar
+java -Xms512m -Xmx2048m -jar -Dwebdriver.chrome.driver="./bin/chromedriver" ./bin/selenium-server-standalone.jar